LRCI-3783 Create a TopLevelBuild interface to hold downstream specific methods
