+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/DefaultTopLevelBuild.java
@@ -8,7 +8,7 @@ package com.liferay.jenkins.results.parser;
 /**
  * @author Peter Yoo
  */
-public class DefaultTopLevelBuild extends TopLevelBuild {
+public class DefaultTopLevelBuild extends BaseTopLevelBuild {
 
 	public DefaultTopLevelBuild(String url) {
 		super(url);
