+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/StandaloneTopLevelBuild.java
@@ -8,7 +8,7 @@ package com.liferay.jenkins.results.parser;
 /**
  * @author Peter Yoo
  */
-public class StandaloneTopLevelBuild extends TopLevelBuild {
+public class StandaloneTopLevelBuild extends BaseTopLevelBuild {
 
 	public StandaloneTopLevelBuild(String url) {
 		super(url);
